review(discount): numbers util
