+++ src/components/DiscountForm/DiscountForm.tsx
@@ -5,7 +5,8 @@ import { useForm } from "react-hook-form";
 import Big from "big.js";
 import { cn } from "@/lib/utils";
 import { useLanguage } from "@/context/language/LanguageContext";
-import { formatDate, parseFloatSafe, parseIntSafe } from "@/utils";
+import { parseFloatSafe, parseIntSafe } from "@/utils/numbers";
+import { formatDate } from "@/utils";
 import { Button } from "../ui/button";
 import { FormattedCurrency } from "../FormattedCurrency";
 import { daysBetween, Act360 } from "./discount-util";
