+++ src/utils/index.ts
@@ -45,15 +45,3 @@ const formatDMY = (date: Date, locale: string): string => {
 export const formatDate = (date: Date, locale: string): string => {
   return formatDMY(date, locale);
 };
-
-export const parseFloatSafe = (str: string | undefined) => {
-  if (str === undefined) return undefined;
-  const parsed = parseFloat(str);
-  return isNaN(parsed) || !isFinite(parsed) ? undefined : parsed;
-};
-
-export const parseIntSafe = (str: string | undefined) => {
-  if (str === undefined) return undefined;
-  const parsed = parseInt(str, 10);
-  return isNaN(parsed) || !isFinite(parsed) ? undefined : parsed;
-};
